+++ Stitch/Graph/Node/View/NodesView.swift
@@ -31,11 +31,11 @@ struct NodesView: View {
             // commentBoxes
             NodesOnlyView(document: document, graph: graph)
             
-            #if DEV || DEV_DEBUG
-            // NOTE: ONLY FOR READING SIZE OF ALL PATCH X NODE-TYPE COMBINATIONS
-            // NEVER CALLED FOR PRODUCTION; AND
-                .modifier(ReadAllPatchAndLayerInputSizes(document: document))
-            #endif
+            //            #if DEV || DEV_DEBUG
+            //            // NOTE: ONLY FOR READING SIZE OF ALL PATCH X NODE-TYPE COMBINATIONS
+            //            // NEVER CALLED FOR PRODUCTION; ONLY CALLED IN DEBUG WHEN E.G. WE HAVE A NEW `Patch`, `NodeType` OR `LayerInputPort`
+            //                .modifier(PSEUDO_SCRIPT_READ_PATCH_NODE_AND_LAYER_INPUT_SIZES(document: document))
+            //            #endif
             
         }
            .modifier(CanvasEdgesViewModifier(document: document, graph: graph))
