+++ Stitch/Graph/Util/HardcodedSizes/HardcodedLayerInputSizes.swift
@@ -1,0 +1,129 @@
+//
+//  HardcodedLayerInputSizes.swift
+//  Stitch
+//
+//  Created by Christian J Clampitt on 5/23/25.
+//
+
+import Foundation
+import SwiftUI
+
+
+// MARK: HARDCODED SWIFT DICTIONARY: REPLACE WHEN WE RUNNING THE PSEUDO-SCRIPT VIEW-MODIFIER AGAIN
+
+let LAYER_INPUT_SIZES_FOR_LAYER_INPUT_PORT: [LayerInputPort: CGSize] = [
+    .orientation: CGSize(width: 318, height: 97),
+    .isShadowsEnabled: CGSize(width: 261, height: 97),
+    .scrollJumpToXStyle: CGSize(width: 324, height: 97),
+    .enabled: CGSize(width: 187, height: 97),
+    .verticalAlignment: CGSize(width: 425, height: 97),
+    .contrast: CGSize(width: 266, height: 97),
+    .cameraDirection: CGSize(width: 352, height: 97),
+    .shape: CGSize(width: 225, height: 97),
+    .scrollYEnabled: CGSize(width: 250, height: 97),
+    .offsetInGroup: CGSize(width: 468, height: 97),
+    .centerAnchor: CGSize(width: 277, height: 97),
+    .sizingScenario: CGSize(width: 277, height: 97),
+    .shadowOpacity: CGSize(width: 318, height: 97),
+    .fitStyle: CGSize(width: 290, height: 97),
+    .fontSize: CGSize(width: 281, height: 97),
+    .itemAlignmentWithinGridCell: CGSize(width: 278, height: 97),
+    .blendMode: CGSize(width: 316, height: 97),
+    .placeholderText: CGSize(width: 261, height: 97),
+    .saturation: CGSize(width: 277, height: 97),
+    .model3D: CGSize(width: 291, height: 97),
+    .brightness: CGSize(width: 280, height: 97),
+    .blur: CGSize(width: 235, height: 97),
+    .scrollJumpToX: CGSize(width: 210, height: 97),
+    .transform3D: CGSize(width: 272, height: 97),
+    .endAngle: CGSize(width: 276, height: 97),
+    .isPinned: CGSize(width: 188, height: 97),
+    .isClipped: CGSize(width: 193, height: 97),
+    .color: CGSize(width: 176, height: 97),
+    .spacingBetweenGridColumns: CGSize(width: 318, height: 97),
+    .pinAnchor: CGSize(width: 253, height: 97),
+    .scrollJumpToXLocation: CGSize(width: 316, height: 97),
+    .strokeWidth: CGSize(width: 295, height: 97),
+    .hueRotation: CGSize(width: 295, height: 97),
+    .maxSize: CGSize(width: 425, height: 97),
+    .textFont: CGSize(width: 385, height: 97),
+    .text: CGSize(width: 210, height: 97),
+    .widthAxis: CGSize(width: 280, height: 97),
+    .scrollXEnabled: CGSize(width: 250, height: 97),
+    .backgroundColor: CGSize(width: 260, height: 97),
+    .endColor: CGSize(width: 226, height: 97),
+    .opacity: CGSize(width: 260, height: 97),
+    .size3D: CGSize(width: 501, height: 97),
+    .rotationX: CGSize(width: 278, height: 97),
+    .isEntityAnimating: CGSize(width: 210, height: 97),
+    .minSize: CGSize(width: 422, height: 97),
+    .spacingBetweenGridRows: CGSize(width: 295, height: 97),
+    .translation3DEnabled: CGSize(width: 215, height: 97),
+    .rotation3DEnabled: CGSize(width: 198, height: 97),
+    .startColor: CGSize(width: 226, height: 97),
+    .zIndex: CGSize(width: 257, height: 97),
+    .cornerRadius: CGSize(width: 303, height: 97),
+    .pinOffset: CGSize(width: 432, height: 97),
+    .canvasLineWidth: CGSize(width: 280, height: 97),
+    .scrollContentSize: CGSize(width: 451, height: 97),
+    .volume: CGSize(width: 258, height: 97),
+    .scrollJumpToYLocation: CGSize(width: 316, height: 97),
+    .strokeLineJoin: CGSize(width: 343, height: 97),
+    .isSwitchToggled: CGSize(width: 220, height: 97),
+    .anchoring: CGSize(width: 249, height: 97),
+    .radius3D: CGSize(width: 254, height: 97),
+    .startAngle: CGSize(width: 283, height: 97),
+    .video: CGSize(width: 285, height: 97),
+    .startAnchor: CGSize(width: 264, height: 97),
+    .isScrollAuto: CGSize(width: 214, height: 97),
+    .contentMode: CGSize(width: 331, height: 97),
+    .position: CGSize(width: 390, height: 97),
+    .strokeColor: CGSize(width: 223, height: 97),
+    .scrollJumpToYStyle: CGSize(width: 324, height: 97),
+    .shadowRadius: CGSize(width: 311, height: 97),
+    .strokePosition: CGSize(width: 336, height: 97),
+    .progressIndicatorStyle: CGSize(width: 270, height: 97),
+    .startRadius: CGSize(width: 289, height: 97),
+    .setupMode: CGSize(width: 222, height: 97),
+    .colorInvert: CGSize(width: 219, height: 97),
+    .scale: CGSize(width: 245, height: 97),
+    .strokeStart: CGSize(width: 287, height: 97),
+    .height3D: CGSize(width: 253, height: 97),
+    .textAlignment: CGSize(width: 371, height: 97),
+    .rotationZ: CGSize(width: 277, height: 97),
+    .videoURL: CGSize(width: 252, height: 97),
+    .mapSpan: CGSize(width: 370, height: 97),
+    .layerMargin: CGSize(width: 738, height: 97),
+    .layerPadding: CGSize(width: 747, height: 97),
+    .deviceAppearance: CGSize(width: 366, height: 97),
+    .textDecoration: CGSize(width: 377, height: 97),
+    .clipped: CGSize(width: 193, height: 97),
+    .strokeEnd: CGSize(width: 280, height: 97),
+    .size: CGSize(width: 394, height: 97),
+    .isMetallic: CGSize(width: 194, height: 97),
+    .anchorEntity: CGSize(width: 250, height: 97),
+    .layerGroupAlignment: CGSize(width: 244, height: 97),
+    .pinTo: CGSize(width: 195, height: 97),
+    .pivot: CGSize(width: 214, height: 97),
+    .strokeLineCap: CGSize(width: 341, height: 97),
+    .endAnchor: CGSize(width: 257, height: 97),
+    .image: CGSize(width: 287, height: 97),
+    .canvasLineColor: CGSize(width: 228, height: 97),
+    .endRadius: CGSize(width: 283, height: 97),
+    .mapLatLong: CGSize(width: 396, height: 97),
+    .scrollJumpToY: CGSize(width: 209, height: 97),
+    .masks: CGSize(width: 184, height: 97),
+    .coordinateSystem: CGSize(width: 363, height: 97),
+    .isAnimating: CGSize(width: 246, height: 97),
+    .heightAxis: CGSize(width: 285, height: 97),
+    .mapType: CGSize(width: 302, height: 97),
+    .shadowOffset: CGSize(width: 436, height: 97),
+    .progress: CGSize(width: 267, height: 97),
+    .isCameraEnabled: CGSize(width: 251, height: 97),
+    .sfSymbol: CGSize(width: 253, height: 97),
+    .shadowColor: CGSize(width: 234, height: 97),
+    .materialThickness: CGSize(width: 290, height: 97),
+    .rotationY: CGSize(width: 277, height: 97),
+    .scale3DEnabled: CGSize(width: 178, height: 97),
+    .spacing: CGSize(width: 263, height: 97),
+]
