+++ Stitch/Graph/Util/HardcodedSizes/HardcodedPatchByNodeTypeSizes.swift
@@ -1,0 +1,175 @@
+//
+//  HardcodedPatchByNodeTypeSizes.swift
+//  Stitch
+//
+//  Created by Christian J Clampitt on 5/23/25.
+//
+
+import Foundation
+import SwiftUI
+
+
+// MARK: HARDCODED SWIFT DICTIONARY: REPLACE WHEN WE RUNNING THE PSEUDO-SCRIPT VIEW-MODIFIER AGAIN
+
+let CANVAS_SIZES_FOR_PATCH_BY_NODE_TYPE: [Patch: [NodeType?: CGSize]] = [
+    .runningTotal: [Optional<NodeType>.none: CGSize(width: 299, height: 97)],
+    .sine: [Optional<NodeType>.none: CGSize(width: 305, height: 97)],
+    .getKeys: [Optional<NodeType>.none: CGSize(width: 242, height: 97)],
+    .arAnchor: [Optional<NodeType>.none: CGSize(width: 387, height: 97)],
+    .equals: [Optional<NodeType>.none: CGSize(width: 293, height: 153)],
+    .arrayJoin: [Optional<NodeType>.none: CGSize(width: 192, height: 125)],
+    .smoothValue: [Optional<NodeType>.none: CGSize(width: 404, height: 153)],
+    .repeatingPulse: [Optional<NodeType>.none: CGSize(width: 297, height: 97)],
+    .convertPosition: [Optional<NodeType>.none: CGSize(width: 530, height: 209)],
+    .add: [.color: CGSize(width: 153, height: 125), .string: CGSize(width: 232, height: 125), .number: CGSize(width: 258, height: 125), .point3D: CGSize(width: 681, height: 125), .size: CGSize(width: 523, height: 125), .position: CGSize(width: 488, height: 125)],
+    .networkRequest: [.string: CGSize(width: 408, height: 237), .media: CGSize(width: 408, height: 237), .json: CGSize(width: 408, height: 237)],
+    .textReplace: [Optional<NodeType>.none: CGSize(width: 298, height: 181)],
+    .jsonObject: [.layerDimension: CGSize(width: 366, height: 125), .padding: CGSize(width: 797, height: 125), .shape: CGSize(width: 328, height: 125), .textFont: CGSize(width: 470, height: 125), .size: CGSize(width: 511, height: 125), .cameraOrientation: CGSize(width: 382, height: 125), .fitStyle: CGSize(width: 382, height: 125), .point3D: CGSize(width: 591, height: 125), .layerStroke: CGSize(width: 382, height: 125), .point4D: CGSize(width: 706, height: 125), .networkRequestType: CGSize(width: 382, height: 125), .string: CGSize(width: 328, height: 125), .transform: CGSize(width: 328, height: 125), .number: CGSize(width: 354, height: 125), .interactionId: CGSize(width: 316, height: 125), .pinToId: CGSize(width: 316, height: 125), .spacing: CGSize(width: 354, height: 125), .anchoring: CGSize(width: 326, height: 125), .color: CGSize(width: 316, height: 125), .bool: CGSize(width: 316, height: 125), .orientation: CGSize(width: 418, height: 125), .animationCurve: CGSize(width: 382, height: 125), .media: CGSize(width: 316, height: 125), .cameraDirection: CGSize(width: 382, height: 125), .textDecoration: CGSize(width: 418, height: 125), .position: CGSize(width: 482, height: 125), .sizingScenario: CGSize(width: 382, height: 125), .json: CGSize(width: 316, height: 125), .pulse: CGSize(width: 316, height: 125), .shapeCommand: CGSize(width: 566, height: 125)],
+    .sampleAndHold: [.size: CGSize(width: 568, height: 153), .layerDimension: CGSize(width: 315, height: 153), .networkRequestType: CGSize(width: 331, height: 153), .fitStyle: CGSize(width: 331, height: 153), .pulse: CGSize(width: 241, height: 153), .textDecoration: CGSize(width: 367, height: 153), .padding: CGSize(width: 1112, height: 153), .color: CGSize(width: 243, height: 153), .position: CGSize(width: 533, height: 153), .anchoring: CGSize(width: 273, height: 153), .cameraOrientation: CGSize(width: 331, height: 153), .textFont: CGSize(width: 419, height: 153), .cameraDirection: CGSize(width: 331, height: 153), .spacing: CGSize(width: 303, height: 153), .orientation: CGSize(width: 367, height: 153), .number: CGSize(width: 303, height: 153), .shapeCommand: CGSize(width: 515, height: 153), .media: CGSize(width: 241, height: 153), .pinToId: CGSize(width: 250, height: 153), .animationCurve: CGSize(width: 331, height: 153), .point3D: CGSize(width: 726, height: 153), .sizingScenario: CGSize(width: 331, height: 153), .transform: CGSize(width: 277, height: 153), .string: CGSize(width: 277, height: 153), .layerStroke: CGSize(width: 331, height: 153), .shape: CGSize(width: 277, height: 153), .json: CGSize(width: 248, height: 153), .point4D: CGSize(width: 929, height: 153), .bool: CGSize(width: 242, height: 153), .interactionId: CGSize(width: 255, height: 153)],
+    .arraySort: [Optional<NodeType>.none: CGSize(width: 269, height: 125)],
+    .not: [Optional<NodeType>.none: CGSize(width: 152, height: 97)],
+    .colorToHex: [Optional<NodeType>.none: CGSize(width: 269, height: 97)],
+    .curveToPack: [Optional<NodeType>.none: CGSize(width: 471, height: 153)],
+    .equalsExactly: [.string: CGSize(width: 219, height: 125), .spacing: CGSize(width: 225, height: 125), .layerDimension: CGSize(width: 229, height: 125), .cameraOrientation: CGSize(width: 245, height: 125), .anchoring: CGSize(width: 220, height: 125), .cameraDirection: CGSize(width: 245, height: 125), .fitStyle: CGSize(width: 245, height: 125), .networkRequestType: CGSize(width: 245, height: 125), .pinToId: CGSize(width: 223, height: 125), .json: CGSize(width: 224, height: 125), .padding: CGSize(width: 661, height: 125), .shapeCommand: CGSize(width: 429, height: 125), .pulse: CGSize(width: 221, height: 125), .size: CGSize(width: 375, height: 125), .sizingScenario: CGSize(width: 245, height: 125), .point3D: CGSize(width: 455, height: 125), .media: CGSize(width: 224, height: 125), .color: CGSize(width: 226, height: 125), .position: CGSize(width: 345, height: 125), .textFont: CGSize(width: 333, height: 125), .point4D: CGSize(width: 569, height: 125), .layerStroke: CGSize(width: 245, height: 125), .textDecoration: CGSize(width: 281, height: 125), .interactionId: CGSize(width: 224, height: 125), .transform: CGSize(width: 223, height: 125), .number: CGSize(width: 223, height: 125), .animationCurve: CGSize(width: 245, height: 125), .bool: CGSize(width: 225, height: 125), .shape: CGSize(width: 223, height: 125), .orientation: CGSize(width: 281, height: 125)],
+    .wirelessBroadcaster: [.cameraOrientation: CGSize(width: 265, height: 97), .bool: CGSize(width: 271, height: 97), .point3D: CGSize(width: 453, height: 97), .textFont: CGSize(width: 332, height: 97), .point4D: CGSize(width: 568, height: 97), .anchoring: CGSize(width: 273, height: 97), .pulse: CGSize(width: 268, height: 97), .position: CGSize(width: 344, height: 97), .networkRequestType: CGSize(width: 270, height: 97), .cameraDirection: CGSize(width: 268, height: 97), .animationCurve: CGSize(width: 270, height: 97), .padding: CGSize(width: 659, height: 97), .layerStroke: CGSize(width: 271, height: 97), .layerDimension: CGSize(width: 267, height: 97), .shape: CGSize(width: 267, height: 97), .fitStyle: CGSize(width: 268, height: 97), .sizingScenario: CGSize(width: 271, height: 97), .spacing: CGSize(width: 264, height: 97), .interactionId: CGSize(width: 266, height: 97), .json: CGSize(width: 268, height: 97), .orientation: CGSize(width: 280, height: 97), .number: CGSize(width: 269, height: 97), .string: CGSize(width: 271, height: 97), .transform: CGSize(width: 269, height: 97), .shapeCommand: CGSize(width: 428, height: 97), .color: CGSize(width: 269, height: 97), .textDecoration: CGSize(width: 280, height: 97), .pinToId: CGSize(width: 263, height: 97), .size: CGSize(width: 373, height: 97), .media: CGSize(width: 269, height: 97)],
+    .loopDedupe: [.shape: CGSize(width: 318, height: 125), .interactionId: CGSize(width: 296, height: 125), .json: CGSize(width: 278, height: 125), .media: CGSize(width: 270, height: 125), .size: CGSize(width: 606, height: 125), .anchoring: CGSize(width: 316, height: 125), .textDecoration: CGSize(width: 408, height: 125), .point4D: CGSize(width: 967, height: 125), .color: CGSize(width: 276, height: 125), .number: CGSize(width: 344, height: 125), .layerDimension: CGSize(width: 356, height: 125), .position: CGSize(width: 571, height: 125), .animationCurve: CGSize(width: 372, height: 125), .transform: CGSize(width: 318, height: 125), .sizingScenario: CGSize(width: 372, height: 125), .bool: CGSize(width: 278, height: 125), .string: CGSize(width: 318, height: 125), .spacing: CGSize(width: 344, height: 125), .fitStyle: CGSize(width: 372, height: 125), .textFont: CGSize(width: 460, height: 125), .point3D: CGSize(width: 764, height: 125), .networkRequestType: CGSize(width: 372, height: 125), .orientation: CGSize(width: 408, height: 125), .pulse: CGSize(width: 278, height: 125), .cameraDirection: CGSize(width: 372, height: 125), .cameraOrientation: CGSize(width: 372, height: 125), .layerStroke: CGSize(width: 372, height: 125), .shapeCommand: CGSize(width: 556, height: 125), .padding: CGSize(width: 1150, height: 125), .pinToId: CGSize(width: 292, height: 125)],
+    .cosine: [Optional<NodeType>.none: CGSize(width: 305, height: 97)],
+    .arrayAppend: [Optional<NodeType>.none: CGSize(width: 294, height: 153)],
+    .reverseProgress: [Optional<NodeType>.none: CGSize(width: 303, height: 153)],
+    .curve: [Optional<NodeType>.none: CGSize(width: 401, height: 125)],
+    .optionSender: [.spacing: CGSize(width: 314, height: 153), .sizingScenario: CGSize(width: 342, height: 153), .textFont: CGSize(width: 430, height: 153), .textDecoration: CGSize(width: 378, height: 153), .pulse: CGSize(width: 271, height: 153), .string: CGSize(width: 311, height: 153), .transform: CGSize(width: 311, height: 153), .color: CGSize(width: 269, height: 153), .layerStroke: CGSize(width: 342, height: 153), .cameraDirection: CGSize(width: 342, height: 153), .networkRequestType: CGSize(width: 342, height: 153), .size: CGSize(width: 579, height: 153), .fitStyle: CGSize(width: 342, height: 153), .interactionId: CGSize(width: 311, height: 153), .point3D: CGSize(width: 737, height: 153), .media: CGSize(width: 263, height: 153), .pinToId: CGSize(width: 311, height: 153), .cameraOrientation: CGSize(width: 342, height: 153), .json: CGSize(width: 309, height: 153), .padding: CGSize(width: 1123, height: 153), .point4D: CGSize(width: 940, height: 153), .anchoring: CGSize(width: 309, height: 153), .number: CGSize(width: 314, height: 153), .animationCurve: CGSize(width: 342, height: 153), .bool: CGSize(width: 271, height: 153), .shapeCommand: CGSize(width: 526, height: 153), .shape: CGSize(width: 311, height: 153), .orientation: CGSize(width: 378, height: 153), .position: CGSize(width: 544, height: 153), .layerDimension: CGSize(width: 326, height: 153)],
+    .mathExpression: [Optional<NodeType>.none: CGSize(width: 237, height: 97)],
+    .loopRemove: [.color: CGSize(width: 348, height: 153), .position: CGSize(width: 571, height: 153), .pulse: CGSize(width: 348, height: 153), .number: CGSize(width: 348, height: 153), .layerDimension: CGSize(width: 356, height: 153), .animationCurve: CGSize(width: 372, height: 153), .layerStroke: CGSize(width: 372, height: 153), .shape: CGSize(width: 348, height: 153), .json: CGSize(width: 348, height: 153), .pinToId: CGSize(width: 348, height: 153), .cameraOrientation: CGSize(width: 372, height: 153), .spacing: CGSize(width: 348, height: 153), .media: CGSize(width: 362, height: 153), .sizingScenario: CGSize(width: 372, height: 153), .interactionId: CGSize(width: 348, height: 153), .anchoring: CGSize(width: 348, height: 153), .padding: CGSize(width: 1150, height: 153), .point3D: CGSize(width: 764, height: 153), .orientation: CGSize(width: 408, height: 153), .textFont: CGSize(width: 460, height: 153), .shapeCommand: CGSize(width: 556, height: 153), .fitStyle: CGSize(width: 372, height: 153), .string: CGSize(width: 348, height: 153), .size: CGSize(width: 606, height: 153), .bool: CGSize(width: 348, height: 153), .cameraDirection: CGSize(width: 372, height: 153), .point4D: CGSize(width: 967, height: 153), .textDecoration: CGSize(width: 408, height: 153), .networkRequestType: CGSize(width: 372, height: 153), .transform: CGSize(width: 348, height: 153)],
+    .loopOverArray: [Optional<NodeType>.none: CGSize(width: 281, height: 125)],
+    .positionUnpack: [Optional<NodeType>.none: CGSize(width: 234, height: 125)],
+    .pack: [.point4D: CGSize(width: 395, height: 125), .point3D: CGSize(width: 395, height: 125), .transform: CGSize(width: 453, height: 125), .position: CGSize(width: 395, height: 125), .size: CGSize(width: 400, height: 125), .shapeCommand: CGSize(width: 577, height: 125)],
+    .mod: [.size: CGSize(width: 523, height: 125), .color: CGSize(width: 157, height: 125), .point3D: CGSize(width: 681, height: 125), .position: CGSize(width: 488, height: 125), .number: CGSize(width: 258, height: 125)],
+    .loopReverse: [.networkRequestType: CGSize(width: 327, height: 97), .anchoring: CGSize(width: 269, height: 97), .shape: CGSize(width: 273, height: 97), .point4D: CGSize(width: 925, height: 97), .transform: CGSize(width: 273, height: 97), .textDecoration: CGSize(width: 363, height: 97), .pulse: CGSize(width: 217, height: 97), .pinToId: CGSize(width: 247, height: 97), .position: CGSize(width: 529, height: 97), .point3D: CGSize(width: 722, height: 97), .interactionId: CGSize(width: 251, height: 97), .media: CGSize(width: 215, height: 97), .spacing: CGSize(width: 299, height: 97), .shapeCommand: CGSize(width: 511, height: 97), .sizingScenario: CGSize(width: 327, height: 97), .cameraDirection: CGSize(width: 327, height: 97), .color: CGSize(width: 214, height: 97), .layerDimension: CGSize(width: 311, height: 97), .orientation: CGSize(width: 363, height: 97), .animationCurve: CGSize(width: 327, height: 97), .json: CGSize(width: 231, height: 97), .textFont: CGSize(width: 415, height: 97), .fitStyle: CGSize(width: 327, height: 97), .string: CGSize(width: 273, height: 97), .number: CGSize(width: 299, height: 97), .size: CGSize(width: 564, height: 97), .bool: CGSize(width: 216, height: 97), .layerStroke: CGSize(width: 327, height: 97), .cameraOrientation: CGSize(width: 327, height: 97), .padding: CGSize(width: 1108, height: 97)],
+    .point4DPack: [Optional<NodeType>.none: CGSize(width: 222, height: 181)],
+    .and: [Optional<NodeType>.none: CGSize(width: 151, height: 125)],
+    .speaker: [Optional<NodeType>.none: CGSize(width: 269, height: 125)],
+    .multiply: [.color: CGSize(width: 178, height: 125), .point3D: CGSize(width: 681, height: 125), .number: CGSize(width: 258, height: 125), .size: CGSize(width: 523, height: 125), .position: CGSize(width: 488, height: 125)],
+    .setValueForKey: [.transform: CGSize(width: 328, height: 153), .json: CGSize(width: 316, height: 153), .pinToId: CGSize(width: 316, height: 153), .animationCurve: CGSize(width: 382, height: 153), .networkRequestType: CGSize(width: 382, height: 153), .orientation: CGSize(width: 418, height: 153), .point3D: CGSize(width: 591, height: 153), .point4D: CGSize(width: 706, height: 153), .interactionId: CGSize(width: 316, height: 153), .spacing: CGSize(width: 354, height: 153), .media: CGSize(width: 316, height: 153), .string: CGSize(width: 328, height: 153), .cameraOrientation: CGSize(width: 382, height: 153), .anchoring: CGSize(width: 326, height: 153), .fitStyle: CGSize(width: 382, height: 153), .shape: CGSize(width: 328, height: 153), .layerDimension: CGSize(width: 366, height: 153), .pulse: CGSize(width: 316, height: 153), .size: CGSize(width: 511, height: 153), .sizingScenario: CGSize(width: 382, height: 153), .bool: CGSize(width: 316, height: 153), .color: CGSize(width: 316, height: 153), .padding: CGSize(width: 797, height: 153), .cameraDirection: CGSize(width: 382, height: 153), .layerStroke: CGSize(width: 382, height: 153), .number: CGSize(width: 354, height: 153), .shapeCommand: CGSize(width: 566, height: 153), .textDecoration: CGSize(width: 418, height: 153), .textFont: CGSize(width: 470, height: 153), .position: CGSize(width: 482, height: 153)],
+    .random: [Optional<NodeType>.none: CGSize(width: 384, height: 153)],
+    .moveToPack: [Optional<NodeType>.none: CGSize(width: 428, height: 97)],
+    .max: [.color: CGSize(width: 155, height: 125), .size: CGSize(width: 523, height: 125), .point3D: CGSize(width: 681, height: 125), .string: CGSize(width: 232, height: 125), .number: CGSize(width: 258, height: 125), .position: CGSize(width: 488, height: 125)],
+    .base64StringToImage: [Optional<NodeType>.none: CGSize(width: 238, height: 97)],
+    .loopBuilder: [.pinToId: CGSize(width: 258, height: 209), .fitStyle: CGSize(width: 338, height: 209), .layerStroke: CGSize(width: 338, height: 209), .cameraDirection: CGSize(width: 338, height: 209), .textFont: CGSize(width: 426, height: 209), .padding: CGSize(width: 1119, height: 209), .size: CGSize(width: 575, height: 209), .string: CGSize(width: 284, height: 209), .cameraOrientation: CGSize(width: 338, height: 209), .bool: CGSize(width: 236, height: 209), .layerDimension: CGSize(width: 322, height: 209), .point3D: CGSize(width: 733, height: 209), .textDecoration: CGSize(width: 374, height: 209), .position: CGSize(width: 540, height: 209), .spacing: CGSize(width: 310, height: 209), .number: CGSize(width: 310, height: 209), .transform: CGSize(width: 284, height: 209), .orientation: CGSize(width: 374, height: 209), .pulse: CGSize(width: 237, height: 209), .point4D: CGSize(width: 936, height: 209), .networkRequestType: CGSize(width: 338, height: 209), .sizingScenario: CGSize(width: 338, height: 209), .shapeCommand: CGSize(width: 522, height: 209), .json: CGSize(width: 242, height: 209), .anchoring: CGSize(width: 280, height: 209), .color: CGSize(width: 235, height: 209), .interactionId: CGSize(width: 262, height: 209), .shape: CGSize(width: 284, height: 209), .animationCurve: CGSize(width: 338, height: 209), .media: CGSize(width: 321, height: 209)],
+    .point4DUnpack: [Optional<NodeType>.none: CGSize(width: 239, height: 181)],
+    .clip: [Optional<NodeType>.none: CGSize(width: 303, height: 153)],
+    .cubicBezierCurve: [Optional<NodeType>.none: CGSize(width: 588, height: 209)],
+    .splitText: [Optional<NodeType>.none: CGSize(width: 279, height: 125)],
+    .valueAtPath: [.padding: CGSize(width: 313, height: 125), .layerStroke: CGSize(width: 313, height: 125), .textDecoration: CGSize(width: 313, height: 125), .json: CGSize(width: 313, height: 125), .spacing: CGSize(width: 313, height: 125), .position: CGSize(width: 313, height: 125), .point3D: CGSize(width: 313, height: 125), .shapeCommand: CGSize(width: 313, height: 125), .bool: CGSize(width: 275, height: 125), .pulse: CGSize(width: 313, height: 125), .cameraOrientation: CGSize(width: 313, height: 125), .textFont: CGSize(width: 313, height: 125), .fitStyle: CGSize(width: 313, height: 125), .media: CGSize(width: 313, height: 125), .orientation: CGSize(width: 313, height: 125), .networkRequestType: CGSize(width: 313, height: 125), .layerDimension: CGSize(width: 313, height: 125), .color: CGSize(width: 313, height: 125), .point4D: CGSize(width: 313, height: 125), .cameraDirection: CGSize(width: 313, height: 125), .animationCurve: CGSize(width: 313, height: 125), .pinToId: CGSize(width: 313, height: 125), .shape: CGSize(width: 313, height: 125), .sizingScenario: CGSize(width: 313, height: 125), .number: CGSize(width: 315, height: 125), .string: CGSize(width: 315, height: 125), .transform: CGSize(width: 313, height: 125), .interactionId: CGSize(width: 313, height: 125), .size: CGSize(width: 313, height: 125), .anchoring: CGSize(width: 313, height: 125)],
+    .optionPicker: [.shape: CGSize(width: 311, height: 153), .media: CGSize(width: 263, height: 153), .layerDimension: CGSize(width: 311, height: 153), .point4D: CGSize(width: 884, height: 153), .interactionId: CGSize(width: 311, height: 153), .sizingScenario: CGSize(width: 311, height: 153), .networkRequestType: CGSize(width: 311, height: 153), .spacing: CGSize(width: 311, height: 153), .orientation: CGSize(width: 322, height: 153), .point3D: CGSize(width: 681, height: 153), .textFont: CGSize(width: 374, height: 153), .padding: CGSize(width: 1067, height: 153), .animationCurve: CGSize(width: 311, height: 153), .cameraOrientation: CGSize(width: 311, height: 153), .cameraDirection: CGSize(width: 311, height: 153), .color: CGSize(width: 269, height: 153), .string: CGSize(width: 311, height: 153), .size: CGSize(width: 523, height: 153), .pinToId: CGSize(width: 311, height: 153), .shapeCommand: CGSize(width: 470, height: 153), .transform: CGSize(width: 311, height: 153), .anchoring: CGSize(width: 309, height: 153), .fitStyle: CGSize(width: 311, height: 153), .number: CGSize(width: 311, height: 153), .json: CGSize(width: 309, height: 153), .pulse: CGSize(width: 271, height: 153), .layerStroke: CGSize(width: 311, height: 153), .bool: CGSize(width: 271, height: 153), .textDecoration: CGSize(width: 322, height: 153), .position: CGSize(width: 488, height: 153)],
+    .indexOf: [Optional<NodeType>.none: CGSize(width: 314, height: 125)],
+    .arcTan2: [Optional<NodeType>.none: CGSize(width: 276, height: 125)],
+    .transition: [.anchoring: CGSize(width: 323, height: 153), .number: CGSize(width: 325, height: 153), .color: CGSize(width: 283, height: 153), .position: CGSize(width: 529, height: 153), .size: CGSize(width: 564, height: 153), .point4D: CGSize(width: 925, height: 153), .point3D: CGSize(width: 722, height: 153)],
+    .min: [.position: CGSize(width: 488, height: 125), .color: CGSize(width: 152, height: 125), .size: CGSize(width: 523, height: 125), .number: CGSize(width: 258, height: 125), .point3D: CGSize(width: 681, height: 125), .string: CGSize(width: 232, height: 125)],
+    .textLength: [Optional<NodeType>.none: CGSize(width: 268, height: 97)],
+    .absoluteValue: [Optional<NodeType>.none: CGSize(width: 258, height: 97)],
+    .loopToArray: [.layerDimension: CGSize(width: 309, height: 97), .fitStyle: CGSize(width: 325, height: 97), .interactionId: CGSize(width: 249, height: 97), .media: CGSize(width: 223, height: 97), .cameraDirection: CGSize(width: 325, height: 97), .json: CGSize(width: 231, height: 97), .bool: CGSize(width: 231, height: 97), .point3D: CGSize(width: 535, height: 97), .shape: CGSize(width: 271, height: 97), .size: CGSize(width: 455, height: 97), .spacing: CGSize(width: 297, height: 97), .networkRequestType: CGSize(width: 325, height: 97), .color: CGSize(width: 229, height: 97), .pulse: CGSize(width: 231, height: 97), .padding: CGSize(width: 741, height: 97), .anchoring: CGSize(width: 269, height: 97), .sizingScenario: CGSize(width: 325, height: 97), .orientation: CGSize(width: 361, height: 97), .position: CGSize(width: 425, height: 97), .shapeCommand: CGSize(width: 509, height: 97), .transform: CGSize(width: 271, height: 97), .animationCurve: CGSize(width: 325, height: 97), .pinToId: CGSize(width: 245, height: 97), .string: CGSize(width: 271, height: 97), .textDecoration: CGSize(width: 361, height: 97), .layerStroke: CGSize(width: 325, height: 97), .point4D: CGSize(width: 649, height: 97), .textFont: CGSize(width: 413, height: 97), .number: CGSize(width: 297, height: 97), .cameraOrientation: CGSize(width: 325, height: 97)],
+    .keyboard: [Optional<NodeType>.none: CGSize(width: 271, height: 97)],
+    .sampleRange: [Optional<NodeType>.none: CGSize(width: 224, height: 153)],
+    .shapeToCommands: [Optional<NodeType>.none: CGSize(width: 367, height: 97)],
+    .dateAndTimeFormatter: [Optional<NodeType>.none: CGSize(width: 343, height: 153)],
+    .qrCodeDetection: [Optional<NodeType>.none: CGSize(width: 442, height: 181)],
+    .splitter: [.cameraOrientation: CGSize(width: 220, height: 97), .color: CGSize(width: 166, height: 97), .orientation: CGSize(width: 256, height: 97), .media: CGSize(width: 210, height: 97), .padding: CGSize(width: 635, height: 97), .bool: CGSize(width: 162, height: 97), .cameraDirection: CGSize(width: 220, height: 97), .point4D: CGSize(width: 544, height: 97), .pulse: CGSize(width: 161, height: 97), .fitStyle: CGSize(width: 220, height: 97), .spacing: CGSize(width: 192, height: 97), .interactionId: CGSize(width: 162, height: 97), .textFont: CGSize(width: 308, height: 97), .number: CGSize(width: 192, height: 97), .string: CGSize(width: 167, height: 97), .json: CGSize(width: 163, height: 97), .sizingScenario: CGSize(width: 220, height: 97), .transform: CGSize(width: 166, height: 97), .layerDimension: CGSize(width: 204, height: 97), .layerStroke: CGSize(width: 220, height: 97), .networkRequestType: CGSize(width: 220, height: 97), .shape: CGSize(width: 166, height: 97), .point3D: CGSize(width: 429, height: 97), .animationCurve: CGSize(width: 220, height: 97), .shapeCommand: CGSize(width: 404, height: 97), .anchoring: CGSize(width: 164, height: 97), .textDecoration: CGSize(width: 256, height: 97), .size: CGSize(width: 349, height: 97), .position: CGSize(width: 320, height: 97), .pinToId: CGSize(width: 165, height: 97)],
+    .pressInteraction: [Optional<NodeType>.none: CGSize(width: 493, height: 237)],
+    .pulseOnChange: [.shape: CGSize(width: 239, height: 97), .shapeCommand: CGSize(width: 475, height: 97), .cameraDirection: CGSize(width: 291, height: 97), .position: CGSize(width: 391, height: 97), .orientation: CGSize(width: 327, height: 97), .layerStroke: CGSize(width: 291, height: 97), .networkRequestType: CGSize(width: 291, height: 97), .interactionId: CGSize(width: 238, height: 97), .color: CGSize(width: 238, height: 97), .textDecoration: CGSize(width: 327, height: 97), .size: CGSize(width: 420, height: 97), .spacing: CGSize(width: 263, height: 97), .anchoring: CGSize(width: 240, height: 97), .animationCurve: CGSize(width: 291, height: 97), .number: CGSize(width: 263, height: 97), .sizingScenario: CGSize(width: 291, height: 97), .string: CGSize(width: 237, height: 97), .point4D: CGSize(width: 615, height: 97), .padding: CGSize(width: 706, height: 97), .layerDimension: CGSize(width: 275, height: 97), .fitStyle: CGSize(width: 291, height: 97), .textFont: CGSize(width: 379, height: 97), .pinToId: CGSize(width: 241, height: 97), .pulse: CGSize(width: 238, height: 97), .transform: CGSize(width: 242, height: 97), .bool: CGSize(width: 242, height: 97), .media: CGSize(width: 239, height: 97), .cameraOrientation: CGSize(width: 291, height: 97), .point3D: CGSize(width: 500, height: 97), .json: CGSize(width: 238, height: 97)],
+    .divide: [.size: CGSize(width: 523, height: 125), .point3D: CGSize(width: 681, height: 125), .position: CGSize(width: 488, height: 125), .number: CGSize(width: 258, height: 125), .color: CGSize(width: 167, height: 125)],
+    .point3DPack: [Optional<NodeType>.none: CGSize(width: 222, height: 153)],
+    .dragInteraction: [Optional<NodeType>.none: CGSize(width: 616, height: 293)],
+    .loop: [Optional<NodeType>.none: CGSize(width: 352, height: 97)],
+    .triangleShape: [Optional<NodeType>.none: CGSize(width: 533, height: 153)],
+    .classicAnimation: [.color: CGSize(width: 292, height: 153), .point4D: CGSize(width: 929, height: 153), .point3D: CGSize(width: 726, height: 153), .anchoring: CGSize(width: 332, height: 153), .position: CGSize(width: 533, height: 153), .size: CGSize(width: 568, height: 153), .number: CGSize(width: 334, height: 153)],
+    .union: [Optional<NodeType>.none: CGSize(width: 232, height: 125)],
+    .arrayCount: [Optional<NodeType>.none: CGSize(width: 236, height: 97)],
+    .time: [Optional<NodeType>.none: CGSize(width: 224, height: 125)],
+    .hapticFeedback: [Optional<NodeType>.none: CGSize(width: 270, height: 125)],
+    .loopCount: [Optional<NodeType>.none: CGSize(width: 299, height: 97)],
+    .cameraFeed: [Optional<NodeType>.none: CGSize(width: 513, height: 153)],
+    .wirelessReceiver: [.number: CGSize(width: 245, height: 97)],
+    .greaterOrEqual: [Optional<NodeType>.none: CGSize(width: 234, height: 125)],
+    .stopwatch: [Optional<NodeType>.none: CGSize(width: 279, height: 153)],
+    .subtract: [.position: CGSize(width: 488, height: 125), .size: CGSize(width: 523, height: 125), .number: CGSize(width: 258, height: 125), .color: CGSize(width: 184, height: 125), .point3D: CGSize(width: 681, height: 125)],
+    .coreMLClassify: [Optional<NodeType>.none: CGSize(width: 425, height: 125)],
+    .loopOptionSwitch: [Optional<NodeType>.none: CGSize(width: 259, height: 97)],
+    .hslColor: [Optional<NodeType>.none: CGSize(width: 293, height: 181)],
+    .squareRoot: [.size: CGSize(width: 523, height: 97), .position: CGSize(width: 488, height: 97), .number: CGSize(width: 258, height: 97), .point3D: CGSize(width: 681, height: 97)],
+    .loopFilter: [.size: CGSize(width: 607, height: 125), .sizingScenario: CGSize(width: 373, height: 125), .number: CGSize(width: 360, height: 125), .fitStyle: CGSize(width: 373, height: 125), .pinToId: CGSize(width: 360, height: 125), .layerDimension: CGSize(width: 360, height: 125), .spacing: CGSize(width: 360, height: 125), .shapeCommand: CGSize(width: 557, height: 125), .layerStroke: CGSize(width: 373, height: 125), .point4D: CGSize(width: 968, height: 125), .transform: CGSize(width: 360, height: 125), .color: CGSize(width: 360, height: 125), .textFont: CGSize(width: 461, height: 125), .interactionId: CGSize(width: 360, height: 125), .string: CGSize(width: 360, height: 125), .media: CGSize(width: 360, height: 125), .cameraDirection: CGSize(width: 373, height: 125), .json: CGSize(width: 360, height: 125), .textDecoration: CGSize(width: 409, height: 125), .anchoring: CGSize(width: 360, height: 125), .position: CGSize(width: 572, height: 125), .animationCurve: CGSize(width: 373, height: 125), .padding: CGSize(width: 1151, height: 125), .point3D: CGSize(width: 765, height: 125), .orientation: CGSize(width: 409, height: 125), .shape: CGSize(width: 360, height: 125), .bool: CGSize(width: 360, height: 125), .networkRequestType: CGSize(width: 373, height: 125), .cameraOrientation: CGSize(width: 373, height: 125), .pulse: CGSize(width: 360, height: 125)],
+    .sizePack: [Optional<NodeType>.none: CGSize(width: 226, height: 125)],
+    .colorToHSL: [Optional<NodeType>.none: CGSize(width: 267, height: 181)],
+    .colorToRGB: [Optional<NodeType>.none: CGSize(width: 238, height: 181)],
+    .flipSwitch: [Optional<NodeType>.none: CGSize(width: 268, height: 153)],
+    .textTransform: [Optional<NodeType>.none: CGSize(width: 361, height: 125)],
+    .sizeUnpack: [Optional<NodeType>.none: CGSize(width: 209, height: 125)],
+    .grayscale: [Optional<NodeType>.none: CGSize(width: 392, height: 97)],
+    .point3DUnpack: [Optional<NodeType>.none: CGSize(width: 239, height: 153)],
+    .coreMLDetection: [Optional<NodeType>.none: CGSize(width: 589, height: 181)],
+    .subarray: [Optional<NodeType>.none: CGSize(width: 390, height: 153)],
+    .length: [.number: CGSize(width: 258, height: 97), .color: CGSize(width: 190, height: 97), .string: CGSize(width: 232, height: 97), .position: CGSize(width: 386, height: 97), .size: CGSize(width: 415, height: 97), .point3D: CGSize(width: 495, height: 97)],
+    .springAnimation: [.point4D: CGSize(width: 929, height: 181), .size: CGSize(width: 568, height: 181), .position: CGSize(width: 533, height: 181), .anchoring: CGSize(width: 324, height: 181), .color: CGSize(width: 284, height: 181), .point3D: CGSize(width: 726, height: 181), .number: CGSize(width: 326, height: 181)],
+    .greaterThan: [Optional<NodeType>.none: CGSize(width: 217, height: 125)],
+    .optionEquals: [.pulse: CGSize(width: 298, height: 153), .point4D: CGSize(width: 622, height: 153), .padding: CGSize(width: 714, height: 153), .shape: CGSize(width: 298, height: 153), .interactionId: CGSize(width: 298, height: 153), .size: CGSize(width: 428, height: 153), .spacing: CGSize(width: 298, height: 153), .textDecoration: CGSize(width: 334, height: 153), .cameraOrientation: CGSize(width: 298, height: 153), .layerDimension: CGSize(width: 298, height: 153), .networkRequestType: CGSize(width: 298, height: 153), .textFont: CGSize(width: 386, height: 153), .media: CGSize(width: 298, height: 153), .transform: CGSize(width: 298, height: 153), .sizingScenario: CGSize(width: 298, height: 153), .cameraDirection: CGSize(width: 298, height: 153), .orientation: CGSize(width: 334, height: 153), .point3D: CGSize(width: 508, height: 153), .color: CGSize(width: 298, height: 153), .number: CGSize(width: 298, height: 153), .position: CGSize(width: 398, height: 153), .bool: CGSize(width: 298, height: 153), .shapeCommand: CGSize(width: 482, height: 153), .json: CGSize(width: 298, height: 153), .fitStyle: CGSize(width: 298, height: 153), .pinToId: CGSize(width: 298, height: 153), .string: CGSize(width: 298, height: 153), .layerStroke: CGSize(width: 298, height: 153), .anchoring: CGSize(width: 298, height: 153), .animationCurve: CGSize(width: 298, height: 153)],
+    .loopSelect: [.spacing: CGSize(width: 384, height: 125), .size: CGSize(width: 607, height: 125), .anchoring: CGSize(width: 384, height: 125), .shape: CGSize(width: 384, height: 125), .point4D: CGSize(width: 968, height: 125), .pinToId: CGSize(width: 384, height: 125), .shapeCommand: CGSize(width: 557, height: 125), .layerDimension: CGSize(width: 384, height: 125), .color: CGSize(width: 384, height: 125), .networkRequestType: CGSize(width: 384, height: 125), .orientation: CGSize(width: 409, height: 125), .animationCurve: CGSize(width: 384, height: 125), .interactionId: CGSize(width: 384, height: 125), .textDecoration: CGSize(width: 409, height: 125), .pulse: CGSize(width: 384, height: 125), .position: CGSize(width: 572, height: 125), .json: CGSize(width: 384, height: 125), .textFont: CGSize(width: 461, height: 125), .transform: CGSize(width: 384, height: 125), .number: CGSize(width: 384, height: 125), .sizingScenario: CGSize(width: 384, height: 125), .fitStyle: CGSize(width: 384, height: 125), .cameraDirection: CGSize(width: 384, height: 125), .point3D: CGSize(width: 765, height: 125), .layerStroke: CGSize(width: 384, height: 125), .bool: CGSize(width: 384, height: 125), .padding: CGSize(width: 1151, height: 125), .string: CGSize(width: 384, height: 125), .cameraOrientation: CGSize(width: 384, height: 125), .media: CGSize(width: 384, height: 125)],
+    .roundedRectangleShape: [Optional<NodeType>.none: CGSize(width: 503, height: 153)],
+    .imageImport: [Optional<NodeType>.none: CGSize(width: 403, height: 125)],
+    .lessThanOrEqual: [Optional<NodeType>.none: CGSize(width: 253, height: 125)],
+    .counter: [Optional<NodeType>.none: CGSize(width: 377, height: 209)],
+    .videoImport: [Optional<NodeType>.none: CGSize(width: 436, height: 209)],
+    .textEndsWith: [Optional<NodeType>.none: CGSize(width: 239, height: 125)],
+    .popAnimation: [.position: CGSize(width: 533, height: 153), .point3D: CGSize(width: 726, height: 153), .anchoring: CGSize(width: 342, height: 153), .color: CGSize(width: 302, height: 153), .number: CGSize(width: 344, height: 153), .size: CGSize(width: 568, height: 153), .point4D: CGSize(width: 929, height: 153)],
+    .loopSum: [Optional<NodeType>.none: CGSize(width: 299, height: 97)],
+    .soundImport: [Optional<NodeType>.none: CGSize(width: 466, height: 237)],
+    .jsonArray: [.media: CGSize(width: 226, height: 125), .json: CGSize(width: 234, height: 125), .networkRequestType: CGSize(width: 328, height: 125), .color: CGSize(width: 232, height: 125), .fitStyle: CGSize(width: 328, height: 125), .spacing: CGSize(width: 300, height: 125), .orientation: CGSize(width: 364, height: 125), .interactionId: CGSize(width: 252, height: 125), .pinToId: CGSize(width: 248, height: 125), .bool: CGSize(width: 234, height: 125), .textFont: CGSize(width: 416, height: 125), .shape: CGSize(width: 274, height: 125), .animationCurve: CGSize(width: 328, height: 125), .cameraDirection: CGSize(width: 328, height: 125), .anchoring: CGSize(width: 272, height: 125), .number: CGSize(width: 300, height: 125), .layerStroke: CGSize(width: 328, height: 125), .cameraOrientation: CGSize(width: 328, height: 125), .padding: CGSize(width: 744, height: 125), .string: CGSize(width: 274, height: 125), .position: CGSize(width: 428, height: 125), .shapeCommand: CGSize(width: 512, height: 125), .point3D: CGSize(width: 538, height: 125), .textDecoration: CGSize(width: 364, height: 125), .layerDimension: CGSize(width: 312, height: 125), .size: CGSize(width: 458, height: 125), .pulse: CGSize(width: 234, height: 125), .point4D: CGSize(width: 652, height: 125), .transform: CGSize(width: 274, height: 125), .sizingScenario: CGSize(width: 328, height: 125)],
+    .unpack: [.point4D: CGSize(width: 628, height: 125), .position: CGSize(width: 404, height: 125), .shapeCommand: CGSize(width: 614, height: 125), .point3D: CGSize(width: 513, height: 125), .transform: CGSize(width: 307, height: 125), .size: CGSize(width: 438, height: 125)],
+    .bouncyConverter: [.number: CGSize(width: 403, height: 125)],
+    .power: [.size: CGSize(width: 523, height: 125), .position: CGSize(width: 488, height: 125), .number: CGSize(width: 258, height: 125), .point3D: CGSize(width: 681, height: 125)],
+    .transformUnpack: [Optional<NodeType>.none: CGSize(width: 250, height: 321)],
+    .closePath: [Optional<NodeType>.none: CGSize(width: 286, height: 97)],
+    .curveToUnpack: [Optional<NodeType>.none: CGSize(width: 1025, height: 153)],
+    .valueAtIndex: [.fitStyle: CGSize(width: 346, height: 125), .orientation: CGSize(width: 346, height: 125), .pulse: CGSize(width: 346, height: 125), .interactionId: CGSize(width: 346, height: 125), .spacing: CGSize(width: 346, height: 125), .textFont: CGSize(width: 346, height: 125), .sizingScenario: CGSize(width: 346, height: 125), .networkRequestType: CGSize(width: 346, height: 125), .transform: CGSize(width: 346, height: 125), .cameraOrientation: CGSize(width: 346, height: 125), .shape: CGSize(width: 346, height: 125), .animationCurve: CGSize(width: 346, height: 125), .layerDimension: CGSize(width: 346, height: 125), .string: CGSize(width: 346, height: 125), .layerStroke: CGSize(width: 346, height: 125), .size: CGSize(width: 346, height: 125), .media: CGSize(width: 346, height: 125), .bool: CGSize(width: 346, height: 125), .shapeCommand: CGSize(width: 346, height: 125), .color: CGSize(width: 346, height: 125), .padding: CGSize(width: 346, height: 125), .point4D: CGSize(width: 346, height: 125), .point3D: CGSize(width: 346, height: 125), .cameraDirection: CGSize(width: 346, height: 125), .json: CGSize(width: 346, height: 125), .position: CGSize(width: 346, height: 125), .pinToId: CGSize(width: 346, height: 125), .textDecoration: CGSize(width: 346, height: 125), .number: CGSize(width: 346, height: 125), .anchoring: CGSize(width: 346, height: 125)],
+    .layerInfo: [Optional<NodeType>.none: CGSize(width: 419, height: 293)],
+    .lineToPack: [Optional<NodeType>.none: CGSize(width: 428, height: 97)],
+    .any: [Optional<NodeType>.none: CGSize(width: 287, height: 125)],
+    .jsonToShape: [Optional<NodeType>.none: CGSize(width: 657, height: 153)],
+    .springFromSettlingDurationAndDampingRatio: [Optional<NodeType>.none: CGSize(width: 447, height: 125)],
+    .progress: [Optional<NodeType>.none: CGSize(width: 303, height: 153)],
+    .whenPrototypeStarts: [Optional<NodeType>.none: CGSize(width: 248, height: 97)],
+    .scrollInteraction: [Optional<NodeType>.none: CGSize(width: 679, height: 461)],
+    .arRaycasting: [Optional<NodeType>.none: CGSize(width: 362, height: 209)],
+    .cubicBezierAnimation: [.number: CGSize(width: 536, height: 237)],
+    .microphone: [Optional<NodeType>.none: CGSize(width: 348, height: 153)],
+    .velocity: [Optional<NodeType>.none: CGSize(width: 258, height: 97)],
+    .imageToBase64String: [Optional<NodeType>.none: CGSize(width: 244, height: 97)],
+    .hexColor: [Optional<NodeType>.none: CGSize(width: 269, height: 97)],
+    .repeatingAnimation: [Optional<NodeType>.none: CGSize(width: 401, height: 209)],
+    .restartPrototype: [Optional<NodeType>.none: CGSize(width: 243, height: 97)],
+    .mouse: [Optional<NodeType>.none: CGSize(width: 338, height: 153)],
+    .positionPack: [Optional<NodeType>.none: CGSize(width: 217, height: 125)],
+    .soulver: [Optional<NodeType>.none: CGSize(width: 266, height: 97)],
+    .loopInsert: [.cameraDirection: CGSize(width: 376, height: 181), .point3D: CGSize(width: 767, height: 181), .spacing: CGSize(width: 348, height: 181), .shape: CGSize(width: 348, height: 181), .networkRequestType: CGSize(width: 376, height: 181), .shapeCommand: CGSize(width: 560, height: 181), .textDecoration: CGSize(width: 412, height: 181), .string: CGSize(width: 348, height: 181), .layerDimension: CGSize(width: 360, height: 181), .point4D: CGSize(width: 970, height: 181), .sizingScenario: CGSize(width: 376, height: 181), .pulse: CGSize(width: 348, height: 181), .position: CGSize(width: 574, height: 181), .animationCurve: CGSize(width: 376, height: 181), .anchoring: CGSize(width: 348, height: 181), .padding: CGSize(width: 1153, height: 181), .media: CGSize(width: 366, height: 181), .pinToId: CGSize(width: 348, height: 181), .json: CGSize(width: 348, height: 181), .bool: CGSize(width: 348, height: 181), .cameraOrientation: CGSize(width: 376, height: 181), .transform: CGSize(width: 348, height: 181), .orientation: CGSize(width: 412, height: 181), .size: CGSize(width: 609, height: 181), .fitStyle: CGSize(width: 376, height: 181), .layerStroke: CGSize(width: 376, height: 181), .number: CGSize(width: 348, height: 181), .interactionId: CGSize(width: 348, height: 181), .color: CGSize(width: 348, height: 181), .textFont: CGSize(width: 464, height: 181)],
+    .deviceInfo: [Optional<NodeType>.none: CGSize(width: 368, height: 265)],
+    .rgba: [Optional<NodeType>.none: CGSize(width: 264, height: 181)],
+    .deviceMotion: [Optional<NodeType>.none: CGSize(width: 452, height: 181)],
+    .circleShape: [Optional<NodeType>.none: CGSize(width: 499, height: 125)],
+    .lessThan: [Optional<NodeType>.none: CGSize(width: 217, height: 125)],
+    .optionSwitch: [Optional<NodeType>.none: CGSize(width: 266, height: 153)],
+    .trimText: [Optional<NodeType>.none: CGSize(width: 320, height: 153)],
+    .deviceTime: [Optional<NodeType>.none: CGSize(width: 266, height: 125)],
+    .delayOne: [Optional<NodeType>.none: CGSize(width: 258, height: 97)],
+    .valueForKey: [.spacing: CGSize(width: 308, height: 125), .point4D: CGSize(width: 308, height: 125), .cameraOrientation: CGSize(width: 308, height: 125), .animationCurve: CGSize(width: 308, height: 125), .string: CGSize(width: 310, height: 125), .padding: CGSize(width: 308, height: 125), .textDecoration: CGSize(width: 308, height: 125), .shapeCommand: CGSize(width: 308, height: 125), .pinToId: CGSize(width: 308, height: 125), .layerStroke: CGSize(width: 308, height: 125), .number: CGSize(width: 310, height: 125), .fitStyle: CGSize(width: 308, height: 125), .layerDimension: CGSize(width: 308, height: 125), .transform: CGSize(width: 308, height: 125), .color: CGSize(width: 308, height: 125), .shape: CGSize(width: 308, height: 125), .interactionId: CGSize(width: 308, height: 125), .bool: CGSize(width: 270, height: 125), .cameraDirection: CGSize(width: 308, height: 125), .networkRequestType: CGSize(width: 308, height: 125), .textFont: CGSize(width: 308, height: 125), .position: CGSize(width: 308, height: 125), .point3D: CGSize(width: 308, height: 125), .anchoring: CGSize(width: 308, height: 125), .pulse: CGSize(width: 308, height: 125), .sizingScenario: CGSize(width: 308, height: 125), .media: CGSize(width: 308, height: 125), .size: CGSize(width: 308, height: 125), .json: CGSize(width: 308, height: 125), .orientation: CGSize(width: 308, height: 125)],
+    .location: [Optional<NodeType>.none: CGSize(width: 373, height: 153)],
+    .round: [Optional<NodeType>.none: CGSize(width: 310, height: 153)],
+    .or: [Optional<NodeType>.none: CGSize(width: 151, height: 125)],
+    .arrayReverse: [Optional<NodeType>.none: CGSize(width: 219, height: 97)],
+    .textStartsWith: [Optional<NodeType>.none: CGSize(width: 238, height: 125)],
+    .loopShuffle: [.cameraOrientation: CGSize(width: 369, height: 125), .textFont: CGSize(width: 457, height: 125), .animationCurve: CGSize(width: 369, height: 125), .color: CGSize(width: 248, height: 125), .padding: CGSize(width: 1150, height: 125), .media: CGSize(width: 242, height: 125), .anchoring: CGSize(width: 311, height: 125), .string: CGSize(width: 315, height: 125), .textDecoration: CGSize(width: 405, height: 125), .pulse: CGSize(width: 250, height: 125), .interactionId: CGSize(width: 293, height: 125), .orientation: CGSize(width: 405, height: 125), .layerDimension: CGSize(width: 353, height: 125), .position: CGSize(width: 571, height: 125), .fitStyle: CGSize(width: 369, height: 125), .point3D: CGSize(width: 764, height: 125), .shape: CGSize(width: 315, height: 125), .pinToId: CGSize(width: 290, height: 125), .transform: CGSize(width: 315, height: 125), .json: CGSize(width: 288, height: 125), .bool: CGSize(width: 249, height: 125), .shapeCommand: CGSize(width: 553, height: 125), .layerStroke: CGSize(width: 369, height: 125), .size: CGSize(width: 606, height: 125), .cameraDirection: CGSize(width: 369, height: 125), .sizingScenario: CGSize(width: 369, height: 125), .point4D: CGSize(width: 967, height: 125), .number: CGSize(width: 341, height: 125), .spacing: CGSize(width: 341, height: 125), .networkRequestType: CGSize(width: 369, height: 125)],
+    .commandsToShape: [Optional<NodeType>.none: CGSize(width: 605, height: 97)],
+    .transformPack: [Optional<NodeType>.none: CGSize(width: 270, height: 321)],
+    .springFromResponseAndDampingRatio: [Optional<NodeType>.none: CGSize(width: 432, height: 125)],
+    .delay: [.number: CGSize(width: 373, height: 153)],
+    .ovalShape: [Optional<NodeType>.none: CGSize(width: 503, height: 125)],
+    .springFromDurationAndBounce: [Optional<NodeType>.none: CGSize(width: 392, height: 125)],
+    .pulse: [Optional<NodeType>.none: CGSize(width: 285, height: 125)],
+]
